Migliorata struttura e Setup Wake Word "Ehi Elia"
