--- config.py
@@ -1,8 +1,0 @@
-from dotenv import load_dotenv
-import os
-load_dotenv()
-
-class Config:
-    ENV = os.getenv("FLASK_ENV", "production")
-    DEBUG = ENV == "development"
-    PORT = int(os.getenv("PORT", "5000"))
