+++ src/elia/server/__init__.py
@@ -1,5 +1,5 @@
 from flask import Flask
-from config import Config
+from elia.config import Config
 
 def create_app():
     app = Flask(__name__, static_folder="server/static", static_url_path="/static")
